pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure_bite"
version = "0.1.4"
description = "A secure authentication and session management system for Django"
readme = "README.MD"
license = {file = "LICENSE"}
authors = [
    {name = "Mbulelo Peyi", email = "notseenyet013@gmail.com"}
]
dependencies = [
    "Django>=3.2",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "django-cors-headers",
    "django-allauth>=0.60.0",
    "dj-rest-auth>=6.0.0",
]
requires-python = ">=3.8"

[tool.setuptools.packages.find]
include = ["secure_bite*"]
exclude = ["secure_bite/tests*"]

[project.urls]
Homepage = "https://github.com/Mbulelo-Peyi/secure_bite"
Repository = "https://github.com/Mbulelo-Peyi/secure_bite"
